first run with gcc14 (#18)
* first run with gcc14

* removed _ALLOC_ definition

* updated RCS logs

* updated ChangeLog

+++ base/subfd.h
@@ -1,5 +1,8 @@
 /*
  * $Log: subfd.h,v $
+ * Revision 1.5  2025-01-21 22:20:26+05:30  Cprogrammer
+ * fixes for gcc14
+ *
  * Revision 1.4  2008-07-14 20:58:52+05:30  Cprogrammer
  * fixed compilation warning on 64 bit os
  *
@@ -21,7 +24,7 @@ extern substdio *subfdout;
 extern substdio *subfdoutsmall;
 extern substdio *subfderr;
 
-ssize_t         subfd_read(int, char *, int);
-ssize_t         subfd_readsmall(int, char *, int);
+ssize_t         subfd_read(int, char *, size_t);
+ssize_t         subfd_readsmall(int, char *, size_t);
 
 #endif
